dialogue-agent: clean up Intent class
Remove/collapse intents that can only be generated from UI buttons.
