+++ lib/dialogue-agent/conversation.js
@@ -16,9 +16,8 @@ const ThingTalk = require('thingtalk');
 
 const ParserClient = require('../prediction/parserclient');
 
-const Semantic = require('./semantic');
-const Intent = Semantic.Intent;
-const ValueCategory = Semantic.ValueCategory;
+const UserInput = require('./user-input');
+const ValueCategory = require('./value-category');
 const Dispatcher = require('./dispatcher');
 
 const DEFAULT_GETTEXT = {
@@ -32,7 +31,7 @@ const DummyStatistics = {
     }
 };
 
-module.exports = class Almond extends events.EventEmitter {
+module.exports = class Conversation extends events.EventEmitter {
     constructor(engine, conversationId, user, delegate, options) {
         super();
         this._engine = engine;
@@ -178,7 +177,7 @@ module.exports = class Almond extends events.EventEmitter {
         }
 
         return this._errorWrap(async () => {
-            const intent = await Intent.parse(root, this.schemas, this._getContext(null, platformData));
+            const intent = await UserInput.parse(root, this.schemas, this._getContext(null, platformData));
             return this._doHandleCommand(intent, null, [], true);
         }, platformData);
     }
@@ -190,7 +189,7 @@ module.exports = class Almond extends events.EventEmitter {
             console.log('Received ThingTalk program');
 
         return this._errorWrap(async () => {
-            const intent = await Intent.parseThingTalk(thingtalk, this.schemas, this._getContext(null, platformData));
+            const intent = await UserInput.parseThingTalk(thingtalk, this.schemas, this._getContext(null, platformData));
             return this._doHandleCommand(intent, null, [], true);
         }, platformData);
     }
@@ -251,13 +250,13 @@ module.exports = class Almond extends events.EventEmitter {
         return Promise.all(analyzed.candidates.map(async (candidate, beamposition) => {
             let parsed;
             try {
-                parsed = await Intent.parse({ code: candidate.code, entities: analyzed.entities }, this.schemas, this._getContext(command, platformData));
+                parsed = await UserInput.parse({ code: candidate.code, entities: analyzed.entities }, this.schemas, this._getContext(command, platformData));
             } catch(e) {
                 // Likely, a type error in the ThingTalk code; not a big deal, but we still log it
                 console.log(`Failed to parse beam ${beamposition}: ${e.message}`);
 
                 if (this._isUnsupportedError(e))
-                    parsed = new Intent.Unsupported(platformData);
+                    parsed = new UserInput.Unsupported(platformData);
                 else
                     return null;
             }
@@ -282,7 +281,7 @@ module.exports = class Almond extends events.EventEmitter {
                 this._lastCandidates = candidates;
 
                 this.stats.hit('sabrina-failure');
-                return this._dispatcher.handle(new Intent.Failed(command, platformData));
+                return this._dispatcher.handle(new UserInput.Failed(command, platformData));
             }
         });
     }
@@ -293,7 +292,7 @@ module.exports = class Almond extends events.EventEmitter {
                 await fn();
             } catch(e) {
                 if (this._isUnsupportedError(e))
-                    await this._doHandleCommand(new Intent.Unsupported(platformData), null, [], true);
+                    await this._doHandleCommand(new UserInput.Unsupported(platformData), null, [], true);
                 else
                     throw e;
             }
@@ -335,10 +334,7 @@ module.exports = class Almond extends events.EventEmitter {
 
         return this._errorWrap(async () => {
             if (this._raw && command !== null) {
-                const value = new ThingTalk.Ast.Value.String(command);
-                const thingtalk = new ThingTalk.Ast.Input.Bookkeeping(null,
-                    new ThingTalk.Ast.BookkeepingIntent.Answer(null, value));
-                let intent = Intent.fromThingTalk(thingtalk, this._getContext(command, platformData));
+                let intent = new UserInput.Answer(new ThingTalk.Ast.Value.String(command), platformData);
                 return this._doHandleCommand(intent, command, [], true);
             }
 
@@ -414,6 +410,3 @@ module.exports = class Almond extends events.EventEmitter {
         return this._delegate.sendResult(message, icon);
     }
 };
-module.exports.Intent = Intent;
-module.exports.ParserClient = ParserClient;
-module.exports.Formatter = ThingTalk.Formatter;
