+++ lib/dialogue-agent/dialogue_policy.js
@@ -11,10 +11,10 @@
 
 const assert = require('assert');
 
-const ValueCategory = require('../semantic').ValueCategory;
-const { CancellationError } = require('../errors');
-const I18n = require('../../i18n');
-const { SentenceGenerator}  = require('../../sentence-generator');
+const ValueCategory = require('./value-category');
+const { CancellationError } = require('./errors');
+const I18n = require('../i18n');
+const { SentenceGenerator}  = require('../sentence-generator');
 
 const MAX_DEPTH = 9;
 const TARGET_PRUNING_SIZE = 50;
