+++ lib/dialogue-agent/dialogue_queue.js
@@ -9,17 +9,12 @@
 // See COPYING for details
 "use strict";
 
-const assert = require('assert');
-const Semantic = require('./semantic');
-const Intent = Semantic.Intent;
-
 class QueueItem {
 }
 
 class UserInput extends QueueItem {
     constructor(intent, confident) {
         super();
-        assert(intent instanceof Intent);
         this.intent = intent;
         this.confident = confident;
     }
