+++ lib/dialogue-agent/dispatcher.js
@@ -17,8 +17,7 @@ const Ast = ThingTalk.Ast;
 const interpolate = require('string-interp');
 const AsyncQueue = require('consumer-queue');
 
-const Semantic = require('./semantic');
-const ValueCategory = Semantic.ValueCategory;
+const ValueCategory = require('./value-category');
 const DialogueExecutor = require('./execution/dialogue_executor');
 const StatementExecutor = require('./execution/statement_executor');
 const QueueItem = require('./dialogue_queue');
@@ -185,11 +184,10 @@ module.exports = class Dispatcher {
         // - Failed when parsing fails
         // - Answer when the user clicks a button, or when the agent is in "raw mode"
         // - NeverMind when the user clicks the X button
-        // - Train when the user clicks/types "train"
         // - Debug when the user clicks/types "debug"
         // - WakeUp when the user says the wake word and nothing else
         if (intent.isFailed) {
-            await getFallbackExamples(this, intent.command);
+            await getFallbackExamples(this, intent.utterance);
             return null;
         }
         if (intent.isUnsupported) {
@@ -202,31 +200,39 @@ module.exports = class Dispatcher {
                 await this.fail();
                 return null;
             }
+            return handled;
         }
 
         if (intent.isProgram) {
-            // handle legacy program intents so we can use "\t" without too much typing
+            // convert thingtalk programs to dialogue states so we can use "\t" without too much typing
             const prediction = new Ast.DialogueState(null, 'org.thingpedia.dialogue.transaction', 'execute', null, []);
             for (let stmt of intent.program.rules)
                  prediction.history.push(new Ast.DialogueHistoryItem(null, stmt, null, 'accepted'));
             return prediction;
         }
 
-        // stop means cancel, but without a failure message
-        if (intent.isStop)
-            throw new CancellationError();
-        if (intent.isNeverMind) {
-            await this.reset();
-            throw new CancellationError();
-        }
+        if (intent.isUICommand) {
+            switch (intent.type) {
+            case 'stop':
+                // stop means cancel, but without a failure message
+                throw new CancellationError();
 
-        if (intent.isDebug) {
-            await this.reply("Current State:\n" + this._dialogueState.prettyprint());
-            return null;
-        }
-        if (intent.isWakeUp) {
-            // nothing to do
-            return null;
+            case 'nevermind':
+                await this.reset();
+                throw new CancellationError();
+
+            case 'debug':
+                await this.reply("Current State:\n" + this._dialogueState.prettyprint());
+                return null;
+
+            case 'wakeup':
+                // nothing to do
+                return null;
+
+            default:
+                await this.fail();
+                return null;
+            }
         }
 
         assert(intent.isDialogueState);
@@ -351,12 +357,6 @@ module.exports = class Dispatcher {
         return queueItem;
     }
 
-    async unexpected() {
-        this.manager.stats.hit('sabrina-unexpected');
-        await this.reply(this._("Sorry, but that's not what I asked."));
-        await this.lookingFor();
-    }
-
     async lookingFor() {
         const ALLOWED_UNITS = {
             'ms': ['ms', 's', 'min', 'h', 'day', 'week', 'mon', 'year'],
@@ -439,17 +439,6 @@ module.exports = class Dispatcher {
         return true;
     }
 
-    async forbid() {
-        await this.reply(this._("I'm sorry, you don't have permission to do that."));
-        await this.setContext(null);
-    }
-    async done() {
-        await this.reply(this._("Consider it done."));
-    }
-    async expect(expected) {
-        this.setExpected(expected);
-        return this.nextIntent();
-    }
     setExpected(expected) {
         if (expected === undefined)
             throw new TypeError();
@@ -460,16 +449,23 @@ module.exports = class Dispatcher {
         this.manager.setContext(context, options);
     }
 
+    /**
+     * Ask a question to the user.
+     *
+     * This is a legacy method used for certain scripted interactions.
+     */
     async ask(expected, question, args) {
         await this.replyInterp(question, args);
-        const intent = await this.expect(expected);
+        await this.setExpected(expected);
+        let intent = await this.nextIntent();
+        while (!intent.isAnswer || intent.category !== expected) {
+            await this.fail();
+            intent = await this.nextIntent();
+        }
         if (expected === ValueCategory.YesNo)
             return intent.value.value;
         else
             return intent.value;
-    }
-    askMoreResults() {
-        return this.expect(ValueCategory.More);
     }
     async askChoices(question, choices) {
         await this.reply(question);
@@ -477,7 +473,12 @@ module.exports = class Dispatcher {
         this._choices = choices;
         for (let i = 0; i < choices.length; i++)
             await this.replyChoice(i, 'choice', choices[i]);
-        return this.nextIntent().then((intent) => intent.value);
+        let intent = await this.nextIntent();
+        while (!intent.isMultipleChoiceAnswer) {
+            await this.fail();
+            intent = await this.nextIntent();
+        }
+        return intent.value;
     }
     async reset() {
         this.manager.stats.hit('sabrina-abort');
