+++ lib/dialogue-agent/entity-linking/contact_search.js
@@ -12,7 +12,7 @@
 const ThingTalk = require('thingtalk');
 const Ast = ThingTalk.Ast;
 
-const ValueCategory = require('../semantic').ValueCategory;
+const ValueCategory = require('../value-category');
 
 // scoring heuristics:
 // prefer full matches, allow prefix matches
