+++ lib/dialogue-agent/entity-linking/user_context.js
@@ -12,7 +12,7 @@
 const ThingTalk = require('thingtalk');
 const Ast = ThingTalk.Ast;
 
-const ValueCategory = require('../semantic').ValueCategory;
+const ValueCategory = require('../value-category');
 const { lookupLocation } = require('./entity_lookup');
 const { tryGetCurrentLocation } = require('../utils');
