--- lib/dialogue-agent/semantic.js
@@ -1,206 +1,0 @@
-// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
-//
-// This file is part of Almond
-//
-// Copyright 2016 The Board of Trustees of the Leland Stanford Junior University
-//
-// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
-//
-// See COPYING for details
-"use strict";
-
-const adt = require('adt');
-
-const ThingTalk = require('thingtalk');
-const Ast = ThingTalk.Ast;
-
-const ValueCategory = adt.data({
-    YesNo: null,
-    MultipleChoice: null,
-
-    Number: null,
-    Measure: { unit: adt.only(String) },
-    RawString: null,
-    Password: null,
-    Date: null,
-    Time: null,
-    Unknown: null,
-    Picture: null,
-    Location: null,
-    PhoneNumber: null,
-    EmailAddress: null,
-    Contact: null,
-    Predicate: null,
-    PermissionResponse: null,
-    Command: null,
-    More: null
-});
-
-ValueCategory.fromType = function fromType(type) {
-    if (type.isEntity && type.type === 'tt:picture')
-        return ValueCategory.Picture;
-    else if (type.isEntity && type.type === 'tt:phone_number')
-        return ValueCategory.PhoneNumber;
-    else if (type.isEntity && type.type === 'tt:email_address')
-        return ValueCategory.EmailAddress;
-    else if (type.isEntity && type.type === 'tt:contact')
-        return ValueCategory.Contact;
-    else if (type.isEntity)
-        return ValueCategory.RawString;
-    else if (type.isBoolean)
-        return ValueCategory.YesNo;
-    else if (type.isString)
-        return ValueCategory.RawString;
-    else if (type.isNumber)
-        return ValueCategory.Number;
-    else if (type.isMeasure)
-        return ValueCategory.Measure(type.unit);
-    else if (type.isEnum)
-        return ValueCategory.RawString;
-    else if (type.isTime)
-        return ValueCategory.Time;
-    else if (type.isDate)
-        return ValueCategory.Date;
-    else if (type.isLocation)
-        return ValueCategory.Location;
-    else
-        return ValueCategory.Unknown;
-};
-
-ValueCategory.fromValue = function fromValue(value) {
-    if (value.isVarRef)
-        return ValueCategory.Unknown;
-
-    var type = value.getType();
-    return ValueCategory.fromType(type);
-};
-
-ValueCategory.toAskSpecial = function toAskSpecial(expected) {
-    let what;
-    if (expected === ValueCategory.YesNo)
-        what = 'yesno';
-    else if (expected === ValueCategory.Location)
-        what = 'location';
-    else if (expected === ValueCategory.Picture)
-        what = 'picture';
-    else if (expected === ValueCategory.PhoneNumber)
-        what = 'phone_number';
-    else if (expected === ValueCategory.EmailAddress)
-        what = 'email_address';
-    else if (expected === ValueCategory.Contact)
-        what = 'contact';
-    else if (expected === ValueCategory.Number)
-        what = 'number';
-    else if (expected === ValueCategory.Date)
-        what = 'date';
-    else if (expected === ValueCategory.Time)
-        what = 'time';
-    else if (expected === ValueCategory.RawString)
-        what = 'raw_string';
-    else if (expected === ValueCategory.Password)
-        what = 'password';
-    else if (expected === ValueCategory.MultipleChoice)
-        what = 'choice';
-    else if (expected === ValueCategory.Command)
-        what = 'command';
-    else if (expected !== null)
-        what = 'generic';
-    else
-        what = null;
-    return what;
-};
-
-const Intent = adt.data({
-    // internally generated intents that have no thingtalk representation
-    Unsupported: { platformData: adt.any },
-    Example: { utterance: adt.only(String), targetCode: adt.only(String), platformData: adt.any },
-    Failed: { command: adt.only(String, null), platformData: adt.any },
-
-    // special UI commands and voice commands that bypass the neural network
-    NeverMind: { thingtalk: adt.only(Ast.Input), platformData: adt.any }, // cancel the current task
-    Stop: { thingtalk: adt.only(Ast.Input), platformData: adt.any }, // cancel the current task, quietly
-    Help: { thingtalk: adt.only(Ast.Input), platformData: adt.any }, // ask for contextual help, or start a new task
-    WakeUp: { thingtalk: adt.only(Ast.Input), platformData: adt.any }, // do nothing and wake up the screen
-    Debug: { thingtalk: adt.only(Ast.Input), platformData: adt.any },
-    Answer: { category: adt.only(ValueCategory), value: adt.only(Ast.Value, Number), thingtalk: adt.only(Ast.Input), platformData: adt.any },
-
-    // thingtalk
-    Program: {
-        program: adt.only(Ast.Program),
-        thingtalk: adt.only(Ast.Input),
-        platformData: adt.any
-    },
-    DialogueState: {
-        prediction: adt.only(Ast.DialogueState),
-        thingtalk: adt.only(Ast.DialogueState),
-        platformData: adt.only
-    }
-});
-
-const SPECIAL_INTENT_MAP = {
-    nevermind: Intent.NeverMind,
-    debug: Intent.Debug,
-    help: Intent.Help,
-    stop: Intent.Stop,
-    wakeup: Intent.WakeUp,
-};
-
-function parseSpecial(thingtalk, context) {
-    let intent;
-    switch (thingtalk.intent.type) {
-    case 'failed':
-        intent = new Intent.Failed(context.command, context.platformData);
-        break;
-    default:
-        if (SPECIAL_INTENT_MAP[thingtalk.intent.type])
-            intent = new (SPECIAL_INTENT_MAP[thingtalk.intent.type])(thingtalk, context.platformData);
-        else
-            intent = new Intent.Failed(context.command, context.platformData);
-    }
-    return intent;
-}
-
-Intent.fromThingTalk = function(thingtalk, context) {
-    if (thingtalk.isBookkeeping) {
-        if (thingtalk.intent.isSpecial)
-            return parseSpecial(thingtalk, context);
-        else if (thingtalk.intent.isAnswer)
-            return new Intent.Answer(ValueCategory.fromValue(thingtalk.intent.value), thingtalk.intent.value, thingtalk, context.platformData);
-        else if (thingtalk.intent.isChoice)
-            return new Intent.Answer(ValueCategory.MultipleChoice, thingtalk.intent.value, thingtalk, context.platformData);
-        else
-            throw new TypeError(`Unrecognized bookkeeping intent`);
-    } else if (thingtalk.isProgram) {
-        return new Intent.Program(thingtalk, thingtalk, context.platformData);
-    } else if (thingtalk.isDialogueState) {
-        return new Intent.DialogueState(thingtalk, thingtalk, context.platformData);
-    } else {
-        throw new TypeError(`Unrecognized ThingTalk command: ${thingtalk.prettyprint()}`);
-    }
-};
-
-Intent.parse = async function parse(json, schemaRetriever, context) {
-    if ('program' in json)
-        return Intent.fromThingTalk(await ThingTalk.Grammar.parseAndTypecheck(json.program, schemaRetriever, true), context);
-
-    let { code, entities } = json;
-    for (let name in entities) {
-        if (name.startsWith('SLOT_')) {
-            let slotname = json.slots[parseInt(name.substring('SLOT_'.length))];
-            let slotType = ThingTalk.Type.fromString(json.slotTypes[slotname]);
-            let value = Ast.Value.fromJSON(slotType, entities[name]);
-            entities[name] = value;
-        }
-    }
-
-    const thingtalk = ThingTalk.NNSyntax.fromNN(code, entities);
-    await thingtalk.typecheck(schemaRetriever, true);
-    return Intent.fromThingTalk(thingtalk, context);
-};
-
-Intent.parseThingTalk = async function parseThingTalk(code, schemaRetriever, context) {
-    return Intent.fromThingTalk(await ThingTalk.Grammar.parseAndTypecheck(code, schemaRetriever, true), context);
-};
-
-module.exports.Intent = Intent;
-module.exports.ValueCategory = ValueCategory;
