+++ lib/dialogue-agent/value-category.js
@@ -1,0 +1,109 @@
+// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
+//
+// This file is part of Genie
+//
+// Copyright 2016-2020 The Board of Trustees of the Leland Stanford Junior University
+//
+// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
+//
+// See COPYING for details
+"use strict";
+
+const adt = require('adt');
+
+const ValueCategory = adt.data({
+    YesNo: null,
+    MultipleChoice: null,
+
+    Number: null,
+    Measure: { unit: adt.only(String) },
+    RawString: null,
+    Password: null,
+    Date: null,
+    Time: null,
+    Unknown: null,
+    Picture: null,
+    Location: null,
+    PhoneNumber: null,
+    EmailAddress: null,
+    Contact: null,
+    Predicate: null,
+    PermissionResponse: null,
+    Command: null,
+    More: null
+});
+module.exports = ValueCategory;
+
+ValueCategory.fromType = function fromType(type) {
+    if (type.isEntity && type.type === 'tt:picture')
+        return ValueCategory.Picture;
+    else if (type.isEntity && type.type === 'tt:phone_number')
+        return ValueCategory.PhoneNumber;
+    else if (type.isEntity && type.type === 'tt:email_address')
+        return ValueCategory.EmailAddress;
+    else if (type.isEntity && type.type === 'tt:contact')
+        return ValueCategory.Contact;
+    else if (type.isEntity)
+        return ValueCategory.RawString;
+    else if (type.isBoolean)
+        return ValueCategory.YesNo;
+    else if (type.isString)
+        return ValueCategory.RawString;
+    else if (type.isNumber)
+        return ValueCategory.Number;
+    else if (type.isMeasure)
+        return ValueCategory.Measure(type.unit);
+    else if (type.isEnum)
+        return ValueCategory.RawString;
+    else if (type.isTime)
+        return ValueCategory.Time;
+    else if (type.isDate)
+        return ValueCategory.Date;
+    else if (type.isLocation)
+        return ValueCategory.Location;
+    else
+        return ValueCategory.Unknown;
+};
+
+ValueCategory.fromValue = function fromValue(value) {
+    if (value.isVarRef)
+        return ValueCategory.Unknown;
+
+    var type = value.getType();
+    return ValueCategory.fromType(type);
+};
+
+ValueCategory.toAskSpecial = function toAskSpecial(expected) {
+    let what;
+    if (expected === ValueCategory.YesNo)
+        what = 'yesno';
+    else if (expected === ValueCategory.Location)
+        what = 'location';
+    else if (expected === ValueCategory.Picture)
+        what = 'picture';
+    else if (expected === ValueCategory.PhoneNumber)
+        what = 'phone_number';
+    else if (expected === ValueCategory.EmailAddress)
+        what = 'email_address';
+    else if (expected === ValueCategory.Contact)
+        what = 'contact';
+    else if (expected === ValueCategory.Number)
+        what = 'number';
+    else if (expected === ValueCategory.Date)
+        what = 'date';
+    else if (expected === ValueCategory.Time)
+        what = 'time';
+    else if (expected === ValueCategory.RawString)
+        what = 'raw_string';
+    else if (expected === ValueCategory.Password)
+        what = 'password';
+    else if (expected === ValueCategory.MultipleChoice)
+        what = 'choice';
+    else if (expected === ValueCategory.Command)
+        what = 'command';
+    else if (expected !== null)
+        what = 'generic';
+    else
+        what = null;
+    return what;
+};
